+++ voting/frontend/app/admin/page.tsx
@@ -3,6 +3,7 @@
 import { useAccount } from "wagmi"
 import { useEffect, useState } from "react"
 import { readContractByFunctionName } from "@/utils"
+import { useRouter } from 'next/navigation';
 
 import VoterManager from "@/components/admin/VoterManager"
 import IsConnected from "@/components/IsConnected";
@@ -10,9 +11,12 @@ import Event from "@/components/Event";
 
 const Admin = () => {
   const { address, isConnected } = useAccount()
+  const { push } = useRouter()
+
   const [owner, setOwner] = useState('')
   const [winningProposalID, setWinningProposalID] = useState<number|null>(null)
   const [workflowStatus, setWorkflowStatus] = useState(0)
+  const [isOwner, setIsOwner] = useState(false)
 
   const WorkflowStatus: string[] = [
     "RegisteringVoters",
@@ -25,10 +29,11 @@ const Admin = () => {
 
   const getOwner = async () => {
     readContractByFunctionName<`0x${string}`>('owner').then(
-        hash => setOwner(hash)
-    ).catch(
-        err => console.log(err.message)
-    )
+      hash => {
+        setOwner(hash)
+        setIsOwner(hash === address)
+        if (hash !== address) push('/')
+    }).catch(() => push('/'))
   }
 
   const getWinningProposalID = async () => {
@@ -53,16 +58,12 @@ const Admin = () => {
       getWinningProposalID()
       getWorkflowStatus()
     }
-  }, [isConnected])
-
-  function isOwner(): boolean {
-    return owner === address;
-  }
+  }, [address, isConnected])
 
   return (
     <div className="flex flex-col space-y-2 mx-auto max-w-screen-lg">
       <IsConnected>
-        {isOwner() ? (
+        {isOwner ? (
           <>
             <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-600 text-indigo-100 shadow-lg">
               <div className="p-4">