+++ voting/frontend/utils/index.ts
@@ -1,6 +1,19 @@
 import { abi, contractAddress } from "@/constants"
+import { Voter } from "@/interfaces/Voter"
 import { readContract, prepareWriteContract, writeContract } from "@wagmi/core"
 import { BaseError, ContractFunctionRevertedError } from "viem"
+
+export const getIsOwner = async (address: `0x${string}`): Promise<boolean> => {
+    return readContractByFunctionName<`0x${string}`>('owner').then(
+        hash => hash === address
+    ).catch(() => false)
+}
+
+export const getIsVoter = async (address: `0x${string}`): Promise<boolean> => {
+    return readContractByFunctionName<Voter>('getVoter', address).then(
+        data => data && data.isRegistered
+    ).catch(() => false)
+}
 
 export const readContractByFunctionName = async <T>(functionName: string, ...args: `0x${string}`[]|string[]): Promise<T> => {
     try {